requirements.txt
fastapi
uvicorn[standard]
python-dotenv
gemini_webapi
tenacity  
cachetools 
aiohttp
python-multipart 